Allow reusing an existing token
The `login` method will send a request to potentially reuse an existing
authorization when the auth config contains the `reuse` field with the value `true`
and the `id` field with the client id.

If no authorization is available, a new one will be returned in the `token` callback
parameter, otherwise the `token` callback parameter will be empty, and the caller
is assumed to know the token.